Fixes for reading the compact trace

+++ rv-predict-source/src/main/java/com/runtimeverification/rvpredict/log/compact/CompactEvent.java
@@ -390,7 +390,7 @@ public abstract class CompactEvent {
         });
     }
 
-    public static List<CompactEvent> signalOutstandingDepth(Context context, long signalDepth) {
+    public static List<CompactEvent> signalOutstandingDepth(Context context, int signalDepth) {
         context.setSignalDepth(signalDepth);
         return NO_EVENTS;
     }
